refactor: rename and relocate UseCase per review comments
Review feedback addressed:
- Move WindowCurrentUrlService from url/ to window/ folder
  (handles window API)
- Rename ApplySavedRulesOnPageLoadUseCase to ApplyRulesOnPageLoadUseCase
- Move UseCase from rule/ to contentOnMessageReceived/ folder
- Update all related imports and test files
