+++ host-frontend-root/frontend-src-root/src/infrastructure/browser/window/WindowCurrentUrlService.ts
