+++ host-frontend-root/frontend-src-root/tests/unit/application/usecases/contentOnMessageReceived/ApplyRulesOnPageLoadUseCase/integration-with-dom-differ.test.ts
@@ -4,12 +4,12 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import type { ICurrentUrlService } from 'src/application/ports/ICurrentUrlService';
 import type { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
-import { ApplySavedRulesOnPageLoadUseCase } from 'src/application/usecases/rule/ApplySavedRulesOnPageLoadUseCase';
+import { ApplyRulesOnPageLoadUseCase } from 'src/application/usecases/contentOnMessageReceived/ApplyRulesOnPageLoadUseCase';
 import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
 import { RewriteRules } from 'src/domain/value-objects/RewriteRules';
 
-describe('ApplySavedRulesOnPageLoadUseCase - DomDiffer Integration', () => {
-  let useCase: ApplySavedRulesOnPageLoadUseCase;
+describe('ApplyRulesOnPageLoadUseCase - DomDiffer Integration', () => {
+  let useCase: ApplyRulesOnPageLoadUseCase;
   let mockRepository: IRewriteRuleRepository;
   let mockCurrentUrlService: ICurrentUrlService;
   let container: HTMLElement;
@@ -20,7 +20,7 @@ describe('ApplySavedRulesOnPageLoadUseCase - DomDiffer Integration', () => {
     mockCurrentUrlService = createMockCurrentUrlService();
 
     // Create usecase instance
-    useCase = new ApplySavedRulesOnPageLoadUseCase(mockRepository, mockCurrentUrlService);
+    useCase = new ApplyRulesOnPageLoadUseCase(mockRepository, mockCurrentUrlService);
 
     // Setup DOM container
     container = document.createElement('div');
